api 컨트롤러와 뷰 컨트롤러 분리
