+++ src/main/java/gift/contoller/ProductController.java
@@ -12,43 +12,35 @@ import gift.dto.ProductResponseDto;
 import gift.service.ProductService;
 import java.util.List;
 
-@Controller
+@RestController
 public class ProductController {
 
 	@Autowired
 	private ProductService productService;
 
-	@GetMapping("/admin")
-	public String viewProducts(Model model) {
-		List<ProductResponseDto> products = productService.getAllProducts().stream()
-			.map(ProductResponseDto::from)
-			.toList();
-		model.addAttribute("products", products);
-		return "admin/mainPage";
-	}
 
-	@PostMapping("/api/product")
+	@PostMapping("/api/v1/product")
 	@ResponseBody
 	public ResponseEntity<?> registerProduct(@RequestBody ProductRequestDto productRequestDto) {
 		productService.saveProduct(productRequestDto);
 		return ResponseEntity.ok().build();
 	}
 
-	@GetMapping("/api/product/{id}")
+	@GetMapping("/api/v1/product/{id}")
 	@ResponseBody
 	public ResponseEntity<ProductResponseDto> getProduct(@PathVariable Long id) {
 		Product product = productService.getProductById(id);
 		return ResponseEntity.ok(ProductResponseDto.from(product));
 	}
 
-	@PutMapping("/api/product/{id}")
+	@PutMapping("/api/v1/product/{id}")
 	@ResponseBody
 	public ResponseEntity<?> updateProduct(@PathVariable Long id, @RequestBody ProductRequestDto productRequestDto) {
 		productService.updateProduct(id, productRequestDto);
 		return ResponseEntity.ok().build();
 	}
 
-	@DeleteMapping("/api/product/{id}")
+	@DeleteMapping("/api/v1/product/{id}")
 	@ResponseBody
 	public ResponseEntity<?> deleteProduct(@PathVariable Long id) {
 		productService.deleteProduct(id);
